fix cmake install; fix pybind util functions

+++ runBuild.sh
@@ -1,18 +1,28 @@
 #!/bin/bash
 
-# install python dev headers
-sudo apt-get install python3-dev
+read -p "Do you want to build or source existing libraries?(b/s) " answer
 
-# install deps
-pip install -r requirementspy.txt
-
-cd python
-
-python3 -m build
-cd dist/
-if python3 -c 'import pkgutil; exit(not pkgutil.find_loader("tool_eventdetectobjects"))'; then
-    echo 'delete existing one'
-    python3 -m pip uninstall tool_eventdetectobjects
+if [[ "$answer" == [yY] ]]; then
+  echo "Yes, continuing..."
+  # install python dev headers
+  sudo apt-get install python3-dev
+  
+  # install deps
+  pip install -r requirementspy.txt
+  
+  cd python
+  
+  python3 -m build
+  cd dist/
+  if python3 -c 'import pkgutil; exit(not pkgutil.find_loader("tool_eventdetectobjects"))'; then
+      echo 'delete existing one'
+      python3 -m pip uninstall tool_eventdetectobjects
+  fi
+  echo 'installing tool_eventdetectobjects...'
+  python3 -m pip install tool_eventdetectobjects-0.0.1-py3-none-any.whl
+else
+  export LD_LIBRARY_PATH=/home/runqiu/code/event_camera_repo/tools/tool_detectobjects/build/tdoinstall/lib:${LD_LIBRARY_PATH}
+  export LD_LIBRARY_PATH=/home/runqiu/code/event_camera_repo/tools/tool_detectobjects/3rdparty/log4cxx/install/lib:${LD_LIBRARY_PATH}
+  echo "LD_LIBRARY_PATH: ${LD_LIBRARY_PATH}"
 fi
-echo 'installing tool_eventdetectobjects...'
-python3 -m pip install tool_eventdetectobjects-0.0.1-py3-none-any.whl
+
